Split env.go into app.go, argument.go and flag.go
