+++ internals/cli/app.go
@@ -3,16 +3,12 @@ package cli
 import (
 	"fmt"
 	"io"
-	"net/url"
 	"os"
-	"strconv"
 	"strings"
 	"text/tabwriter"
-	"time"
 
 	"bitbucket.org/zombiezen/cardcpx/natsort"
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 )
 
 var (
@@ -332,235 +328,6 @@ func (c *CommandClause) BindAction(fn func() error) {
 			return fn()
 		}
 	}
-}
-
-type FlagSet struct {
-	*pflag.FlagSet
-	cmd *CommandClause
-}
-
-func (f *FlagSet) BoolVarP(reference *bool, name, shorthand string, def bool, usage string) *Flag {
-	f.FlagSet.BoolVarP(reference, name, shorthand, def, usage)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) IntVarP(reference *int, name, shorthand string, def int, usage string) *Flag {
-	f.FlagSet.IntVarP(reference, name, shorthand, def, usage)
-	f.cmd.Flag(name).DefValue = strconv.Itoa(def)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) StringVarP(reference *string, name, shorthand string, def string, usage string) *Flag {
-	f.FlagSet.StringVarP(reference, name, shorthand, def, usage)
-	f.cmd.Flag(name).DefValue = def
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) DurationVarP(reference *time.Duration, name, shorthand string, def time.Duration, usage string) *Flag {
-	f.FlagSet.DurationVarP(reference, name, shorthand, def, usage)
-	f.cmd.Flag(name).DefValue = shortDur(reference)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) BoolVar(reference *bool, name string, def bool, usage string) *Flag {
-	f.FlagSet.BoolVar(reference, name, def, usage)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) IntVar(reference *int, name string, def int, usage string) *Flag {
-	f.FlagSet.IntVar(reference, name, def, usage)
-	f.cmd.Flag(name).DefValue = strconv.Itoa(def)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) StringVar(reference *string, name string, def string, usage string) *Flag {
-	f.FlagSet.StringVar(reference, name, def, usage)
-	f.cmd.Flag(name).DefValue = def
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) DurationVar(reference *time.Duration, name string, def time.Duration, usage string) *Flag {
-	f.FlagSet.DurationVar(reference, name, def, usage)
-	f.cmd.Flag(name).DefValue = shortDur(reference)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) VarP(reference pflag.Value, name string, shorthand string, usage string) *Flag {
-	f.FlagSet.VarP(reference, name, shorthand, usage)
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) Var(reference pflag.Value, name string, usage string) *Flag {
-	f.FlagSet.Var(reference, name, usage)
-	f.cmd.Flag(name).DefValue = reference.String()
-	return f.cmd.Flag(name)
-}
-
-func (f *FlagSet) VarPF(reference pflag.Value, name string, shorthand string, usage string) *pflag.Flag {
-	flag := f.FlagSet.VarPF(reference, name, shorthand, usage)
-	f.cmd.Flag(name)
-	return flag
-}
-
-// Flag represents a command-line flag.
-type Flag struct {
-	*pflag.Flag
-
-	envVar string
-	app    *App
-}
-
-// Envar overrides the environment variable name that configures the default
-// value for a flag.
-func (f *Flag) Envar(name string) *Flag {
-	name = strings.ToUpper(name)
-	if f.envVar != "" {
-		f.app.unregisterEnvVar(f.envVar)
-	}
-	f.app.registerEnvVar(name)
-	f.envVar = name
-	if os.Getenv(f.envVar) != "" {
-		f.Flag.DefValue = os.Getenv(f.envVar)
-	}
-	return f
-}
-
-func (f *Flag) HasEnvarValue() bool {
-	return os.Getenv(f.envVar) != ""
-}
-
-func (f *Flag) NoEnvar() *Flag {
-	if f.envVar != "" {
-		f.app.unregisterEnvVar(f.envVar)
-	}
-	f.envVar = ""
-	return f
-}
-
-// formatName takes a name and converts it to an uppercased name,
-// joined by the given separator and prefixed with the given prefix.
-func formatName(name, prefix, separator string, delimiters ...string) string {
-	if name == "" {
-		return strings.ToUpper(prefix)
-	}
-	for _, delim := range delimiters {
-		name = strings.Replace(name, delim, separator, -1)
-	}
-
-	if prefix == "" {
-		return strings.ToUpper(name)
-	}
-
-	return strings.ToUpper(strings.Join([]string{prefix, name}, separator))
-}
-
-// splitVar gets the key value pair from a key=value declaration, returning
-// true if it matches the given prefix.
-func splitVar(prefix, separator, envVar string) (string, string, bool) {
-	envVar = strings.TrimSpace(envVar)
-	split := strings.Split(envVar, "=")
-	if len(split) != 2 {
-		return "", "", false
-	}
-
-	prefix = fmt.Sprintf("%s%s", strings.ToUpper(prefix), separator)
-	return split[0], split[1], strings.HasPrefix(strings.ToUpper(split[0]), prefix)
-}
-
-type Argument struct {
-	Value       ArgValue
-	Name        string
-	Required    bool
-	Placeholder string
-	Description string
-	Hidden      bool
-}
-
-type ArgValue interface {
-	Set(string) error
-}
-
-func ArgumentRegister(params []Argument, args []string) error {
-	for i, arg := range args {
-		err := params[i].Value.Set(arg)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func ArgumentArrRegister(params []Argument, args []string) error {
-	for _, arg := range args {
-		err := params[0].Value.Set(arg)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-type StringValue struct {
-	Param string
-}
-
-func (s *StringValue) Set(replacer string) error {
-	s.Param = replacer
-	return nil
-}
-
-type StringArrValue struct {
-	Param []string
-}
-
-func (s *StringArrValue) Set(replacer string) error {
-	s.Param = append(s.Param, replacer)
-	return nil
-}
-
-type URLValue struct {
-	*url.URL
-}
-
-func (s *URLValue) Set(replacer string) error {
-	var err error
-	s.URL, err = url.Parse(replacer)
-	return err
-}
-
-type ByteValue struct {
-	Param []byte
-}
-
-func (s *ByteValue) Set(replacer string) error {
-	s.Param = []byte(replacer)
-	return nil
-}
-
-// Registerer allows others to register commands on it.
-type Registerer interface {
-	Command(cmd string, help string) *CommandClause
-}
-
-func getRequired(params []Argument) int {
-	required := 0
-	for _, arg := range params {
-		if arg.Required {
-			required++
-		}
-	}
-	return required
-}
-
-func shortDur(d *time.Duration) string {
-	s := d.String()
-	if strings.HasSuffix(s, "m0s") {
-		s = s[:len(s)-2]
-	}
-	if strings.HasSuffix(s, "h0m") {
-		s = s[:len(s)-2]
-	}
-	return s
 }
 
 type preRunAction func(*cobra.Command, []string) error
